Matlab code
The tester is asked to determine if there is a "target" (the object possessing an unique feature) among a set of randomly displayed characters. They should press T or F on their keyboards to make judgement  as soon as possible. Two different fashion of displaying objects are adopted, each corresponds to a particular paradigm of visual search (feature search and conjunction search). The Pearson Correlation Coefficients for each type of search are reported after the experiment.